refactor: separate guild models into models.py
Moved Pydantic models (GuildInfo, GuildDetails) from endpoints.py
to a dedicated models.py file for better code organization.

This follows the same pattern used in other server endpoints like
links, roles, etc.
